Dev (#2)
* Implement food management features with FoodController, FoodService, and FoodRepository; add Food and User models; configure application properties and SQL data initialization.

* Started Implementing SpringSecurity

* Refactor FoodController and FoodService to return ResponseEntity; remove UserService

* Remove User and Security Implementation and improved Food Controller and Service

+++ src/main/java/com/petcc/restaurante/service/FoodService.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.petcc.restaurante.model.Food;
 import com.petcc.restaurante.repository.FoodRepository;
 
@@ -15,17 +14,22 @@ public class FoodService {
     public List<Food> getAllFoods(){
         return foodRepository.findAll();
     }
-    public void saveFood(String title, double price, String image){
-        foodRepository.save(new Food(title, price, image));
+    public Food saveFood(Food food){
+        return foodRepository.save(food);
+       
     }
-    public void deleteFood(Long id){
-        foodRepository.deleteById(id);
+    public boolean deleteFood(Long id){
+        if (foodRepository.existsById(id)) {
+            foodRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
-    public void updateFood(Long id, String title, double price, String image){
+    public Food updateFood(Long id, String title, double price, String image){
         Food food = foodRepository.findById(id).orElseThrow(() -> new RuntimeException("Food not found"));
         food.setTitle(title);
         food.setPrice(price);
         food.setImage(image);
-        foodRepository.save(food);
+        return foodRepository.save(food);
     }
 }
